// src/modules/auth/jwt-payload.interface.ts
export interface IJwtPayload {
  usu_id: number;
  usu_nombre: string;
  usu_alias: string;
  usu_apellido: string;
  usu_activo: boolean;
  emp_id: number;
  usu_administrador: boolean;
  pve_id: number;
  usu_contrasena: string;
  reg_id: number;
  puntoventa: string;
  almacen: string;
  region: string;
}